# Change dns_tolower() and dns_toupper() to use a table