--- a/pdns/rcpgenerator.cc
+++ b/pdns/rcpgenerator.cc
@@ -23,7 +23,6 @@
 #include "config.h"
 #endif
 #include "rcpgenerator.hh"
-#include "ascii.hh"
 #include "dnsparser.hh"
 #include "misc.hh"
 #include "utility.hh"