… pdns/ascii.hh
… pdns/dnsdistdist/ascii.hh
… pdns/dnsname.cc
… pdns/recursordist/ascii.hh