### Short description
<!-- Write a small description of what this Pull Request fixes or provides, including the issue #s -->

I figured this micro-optimization should lead to a speedup as there are less conditionals, smaller code and the page containing the tables will be cached almost certainly. And indeed:

New speedtest:
```
    Before:
    'BurtleHashCI' 0.10 seconds: 66975796.8 runs/s, 0.01 usec/run
    
    After:
    'BurtleHashCI' 0.10 seconds: 96758136.6 runs/s, 0.01 usec/run
```
I'm also folding `ascii.hh` into `dnsname.{cc,hh}

### Checklist
<!-- please indicate if any of these things are done/included with this Pull Request. Not all boxes need to be checked for the Pull Request to be accepted -->
I have:
- [X] read the [CONTRIBUTING.md](https://github.com/PowerDNS/pdns/blob/master/CONTRIBUTING.md) document
- [X] compiled this code
- [X] tested this code
- [ ] included documentation (including possible behaviour changes)
- [ ] documented the code
- [ ] added or modified regression test(s)
- [X] added or modified unit test(s)